Separate the pool-details hook for smart contract and database.
